requirements.txt
# FastAPI e dependências core
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validação e configuração
pydantic[email]==2.5.0
python-decouple==3.8

# Utilitários
python-dateutil==2.8.2
pytz==2023.3

# Email
aiosmtplib==3.0.1

# Logging
structlog==23.2.0

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Documentação
mkdocs==1.5.3
mkdocs-material==9.4.8

# Linting e formatação
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoramento (opcional)
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Para desenvolvimento local
python-dotenv==1.0.0